requirements/dev.txt
-r base.txt
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
requests-mock>=1.11.0
pytest-mock>=3.14.1
